Add debuginfo to reduction `.expand` loops
Add debug symbols necessary for tracking general reduction kernel
invocations on a particular co-ordinate in the debugger. The list of
exported general reduce kernels is fetched from
`bcinfo::MetadataExtractor`, and the <accumulator>.expand
is added to the list of functions to be complemented with debug
metadata.

This works because reduction accumulators are expanded using the
same logic as ForEach kernels in `bcc::RSKernelExpandPass`

Test: Pending testuite to be pushed to frameworks/rs/tests/lldb in a
subsequent request.

Change-Id: I54f216fab95213fe2b070ed81e9e05c738943a9d
Signed-off-by: Eliminater74 <eliminater74@gmail.com>